+++ apparat_launcher/plugin_passwordgen.py
@@ -25,7 +25,7 @@ TRIGGER = ('!pw', '!password',)
 # -----------------------------------------------------------------------------------------------
 def prepare_general(current_search_string, main_window):
     """Prepare General"""
-    tools.debug_output(__name__, 'prepare_general', 'starting', 1,)
+    tools.debug_output(__name__, 'prepare_general', 'starting', 1)
 
     ## Reset status notification back to OK
     main_window.status_notification_reset()
@@ -33,19 +33,19 @@ def prepare_general(current_search_string, main_window):
     icon_size = ini.read_single_ini_value('General', 'icon_size') # get preference value
 
     if current_search_string == ('!pw') or current_search_string == ('!password'):
-        tools.debug_output(__name__, 'prepare_general', 'Case: Password Generator', 1,)
+        tools.debug_output(__name__, 'prepare_general', 'Case: Password Generator', 1)
         prepare_plugin_passwordgen(main_window, icon_size)
 
     else:
-        tools.debug_output(__name__, 'prepare_general', 'Error: Unexpected passwordgen plugin command', 3,)
+        tools.debug_output(__name__, 'prepare_general', 'Error: Unexpected passwordgen plugin command', 3)
         main_window.status_notification_display_error('Unexpected passwordgen plugin command')
 
-    tools.debug_output(__name__, 'prepare_general', 'finished', 1,)
+    tools.debug_output(__name__, 'prepare_general', 'finished', 1)
 
 
 def prepare_plugin_passwordgen(main_window, icon_size):
     """Prepares UI for plugin PasswordGen"""
-    tools.debug_output(__name__, 'prepare_plugin_passwordgen', 'starting', 1,)
+    tools.debug_output(__name__, 'prepare_plugin_passwordgen', 'starting', 1)
     main_window.plugin__update_general_ui_information('Password Generator') ## update plugin info
 
     ## command button & txt
@@ -58,16 +58,16 @@ def prepare_plugin_passwordgen(main_window, icon_size):
     main_window.ui__bt_parameter_img = wx.Image('gfx/core/'+icon_size+'/execute.png', wx.BITMAP_TYPE_PNG)
     main_window.ui__bt_parameter.SetBitmap(main_window.ui__bt_parameter_img.ConvertToBitmap())
 
-    tools.debug_output(__name__, 'prepare_plugin_passwordgen', 'Finished preparing password UI', 1,)
+    tools.debug_output(__name__, 'prepare_plugin_passwordgen', 'Finished preparing password UI', 1)
 
 
 def execute_password_generation(main_window):
     """Handles the actual password generation process"""
-    tools.debug_output(__name__, 'execute_password_generation', 'started', 0,)
+    tools.debug_output(__name__, 'execute_password_generation', 'started', 1)
     dlg = wx.TextEntryDialog(None, 'Please insert desired password length', 'Password Generator', '8')
     ret = dlg.ShowModal()
     if ret == wx.ID_OK:
-        tools.debug_output(__name__, 'execute_password_generation', 'Password length set to: '+dlg.GetValue(), 1,)
+        tools.debug_output(__name__, 'execute_password_generation', 'Password length set to: '+dlg.GetValue(), 1)
         try:
             password_length = int(dlg.GetValue())
             if password_length < 8:
@@ -81,7 +81,7 @@ def execute_password_generation(main_window):
             generated_passwords = ''
 
             if pw_type == wx.ID_YES:
-                tools.debug_output(__name__, 'execute_password_generation', 'User selected memorable password type', 1,)
+                tools.debug_output(__name__, 'execute_password_generation', 'User selected memorable password type', 1)
 
                 for x in range(0, 5): # generate 5 memorizable passwords
                     tools.debug_output(__name__, 'execute_password_generation', 'Generating memorable password '+str(x), 1)
@@ -94,10 +94,10 @@ def execute_password_generation(main_window):
                 generated_passwords = generated_passwords+'\n\nXKCD (936) like:\n'+xkcd
 
             else:
-                tools.debug_output(__name__, 'execute_password_generation', 'User selected default password type', 1,)
+                tools.debug_output(__name__, 'execute_password_generation', 'User selected default password type', 1)
 
                 for x in range(0, 5): # generate 5 default passwords
-                    tools.debug_output(__name__, 'execute_password_generation', 'Generating general password '+str(x), 1,)
+                    tools.debug_output(__name__, 'execute_password_generation', 'Generating general password '+str(x), 1)
 
                     chars = string.ascii_letters + string.digits + '!@#$%^&*()'
                     random.seed = (os.urandom(1024))
@@ -109,16 +109,16 @@ def execute_password_generation(main_window):
             wx.MessageBox('Choose from the following:\n\n'+generated_passwords, 'Password Generator', wx.OK | wx.ICON_INFORMATION)
 
             ## update usage-statistics
-            tools.debug_output(__name__, 'execute_password_generation', 'Updating statistics (plugin_executed)', 1,)
+            tools.debug_output(__name__, 'execute_password_generation', 'Updating statistics (plugin_executed)', 1)
             current_plugin_executed_count = ini.read_single_ini_value('Statistics', 'plugin_executed') # get current value from ini
             ini.write_single_ini_value('Statistics', 'plugin_executed', int(current_plugin_executed_count)+1) # update ini +1
 
         except ValueError:
-            tools.debug_output(__name__, 'execute', 'Password length entered by user was not a number', 3,)
+            tools.debug_output(__name__, 'execute', 'Password length entered by user was not a number', 3)
             wx.MessageBox('Length was not a number', 'Password Generator', wx.OK | wx.ICON_WARNING)
 
     else:
-        tools.debug_output(__name__, 'execute', 'Password length definition canceld by user', 2,)
+        tools.debug_output(__name__, 'execute', 'Password length definition canceld by user', 2)
 
     ## reset the UI
     main_window.reset_ui()
@@ -162,5 +162,5 @@ def generate_xkcd_password():
         single = get_random_word()
         single = single.rstrip()
         full = full+' '+single
-    tools.debug_output(__name__, 'generate_xkcd_password', 'Finished generating an xkcd-like password', 1,)
+    tools.debug_output(__name__, 'generate_xkcd_password', 'Finished generating an xkcd-like password', 1)
     return full